requirements.txt
aioredis==1.3.1
amqp==5.0.5
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
billiard==3.6.3.0
celery==5.0.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
Django==3.1.6
djangorestframework==3.12.2
hiredis==1.1.0
hyperlink==21.0.0
idna==3.1
importlib-metadata==3.7.0
incremental==17.5.0
kombu==5.0.2
msgpack==1.0.2
Pillow==8.1.0
pkg-resources==0.0.0
prompt-toolkit==3.0.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyotp==2.6.0
pytz==2021.1
redis==3.5.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
zope.interface==5.2.0
